+++ packages/wonder-blocks-dropdown/src/components/__tests__/multi-select.test.tsx
@@ -213,8 +213,8 @@ describe("MultiSelect", () => {
                 `${numSelectedValues} fruits`,
         };
 
-        const ControlledComponent: React.FC<Props> = function (
-            props,
+        const ControlledComponent = function (
+            props: Props,
         ): React.ReactElement {
             const [opened, setOpened] = React.useState(props.opened ?? false);
             const [selectedValues, setSelectedValues] = React.useState([]);